pochtiprintf

+++ ft_printf/do_char.c
@@ -6,7 +6,7 @@
 /*   By: rloraine <rloraine@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/29 14:26:03 by rloraine          #+#    #+#             */
-/*   Updated: 2019/07/04 16:49:37 by rloraine         ###   ########.fr       */
+/*   Updated: 2019/07/09 18:01:58 by rloraine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 int		do_c_wm(const wchar_t c, t_format *prms)
 {
 	prms->len = 1;
-	if (!(prms->flag & MINUS) && prms->width > prms->len)
+	if (!(prms->flag & MINUS) && (int)prms->width > (int)prms->len)
 		char_to_buf((prms->flag & ZERO ? '0' : ' '), prms->width - prms->len);
 	if (prms->spec == 'c' && prms->mod != L)
 		char_to_buf(c, 1);
-	if ((prms->flag & MINUS) && prms->width > prms->len)
+	if ((prms->flag & MINUS) && (int)prms->width > (int)prms->len)
 		char_to_buf((prms->flag & ZERO ? '0' : ' '), prms->width - prms->len);
 	return (0);
 }
@@ -32,7 +32,7 @@ int		do_s_wm(const char *str, t_format *params)
 		str = (params->acc ? "(null)" : "");
 	params->len = ft_strlen(str);
 	if (params->flag & ACC)
-		params->len = params->acc < params->len ? params->acc : params->len;
+		params->len = (int)params->acc < (int)params->len ? params->acc : params->len;
 	width = params->width - params->len;
 	if (!(params->flag & MINUS) && width > 0)
 		char_to_buf((params->flag & ZERO ? '0' : ' '), width);
